+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2022-12-08 -- v0.0.2
+### Added
+- Initial commit of Avro encoder, S3 client, and Kinesis client
+- Initial commit of main method
+### Fixed
+- Re-formatted Sierra and Geocoder API clients to follow new linter guidelines
+
 ## 2022-11-29 -- v0.0.1
 ### Added
-- Initial commit
+- Initial commit of Sierra and Geocoder API clients